+++ pyutil/genetic_utils.py
@@ -1,13 +1,3 @@
-"""
-# genetic algorithm notes
-- look into simulated annealing
-- make a "lean" version with:
-	- no printing of position/activation @ every timestep, only print at end
-	- dont print anything except the final position to stdout
-"""
-
-from genericpath import isdir
-from posixpath import join
 from typing import *
 import subprocess
 import copy
@@ -20,398 +10,16 @@ import numpy as np # type: ignore
 from nptyping import NDArray # type: ignore
 from pydbg import dbg # type: ignore
 
-if TYPE_CHECKING:
-	from mypy_extensions import Arg
-else:
-	Arg = lambda t,s : t
-
-from pyutil.util import (
-	ModParam, ModParamsDists, ModTypes, NormalDistTuple, Path,mkdir,joinPath,
-	strList_to_dict,ParamsDict,ModParamsDict,ModParamsRanges,
-	RangeTuple,norm_prob,distributions_to_ranges,
-	VecXY,dump_state,
-	find_conn_idx,find_conn_idx_regex,
-	genCmd_singlerun,
-	dict_hash,load_params,dict_to_filename,modprmdict_to_filename,
-	keylist_access_nested_dict,
-	read_body_data,CoordsRotArr,
-	prntmsg,
-)
-
+from pyutil.util import *
+from pyutil.params import *
+from pyutil.extract_run_data import *
 from pyutil.geno_distr import DEFAULT_DISTS,DEFAULT_EVALRUNS
-
 
 Process = Any
 Population = List[ModParamsDict]
 PopulationFitness = List[
 	Tuple[ModParamsDict, float]
 ]
-
-
-
-"""
-
- #    # ###### #####   ####  ######
- ##  ## #      #    # #    # #
- # ## # #####  #    # #      #####
- #    # #      #####  #  ### #
- #    # #      #   #  #    # #
- #    # ###### #    #  ####  ######
-
-"""
-
-def merge_params_with_mods(
-		# base params
-		params_base : ParamsDict,
-		# modified params (this is what we are optimizing)
-		params_mod : ModParamsDict,
-	) -> ParamsDict:
-	"""merges a params file with a special "mod" dict
-	
-	returns a modified copy of `params_base`, modified according to the contents of `params_mod`
-	`params_base` is of the same form as a regular params.json file,
-	but `params_mod` has the following structure:
-
-	```python
-	params_mod = {
-			('params','Head.neurons.AWA.theta' : 2.0,
-			('params','ChemoReceptors.alpha' : 2.0,
-			('conn','Head,AWA,RIM,chem' : 10.0,
-		}
-	}
-	```
-
-	- keys starting with `params` map dot-separated keys to the nested params dict, to their desired values
-	- keys starting with `conn` map comma-separated connection identifiers to their desired values
-	
-	### Parameters:
-	 - `params_mod : ModParamsDict` 
-	   special dict to modify a copy of `params_base`
-	 - `params_base : ParamsDict`
-	   `params.json` style dict
-	
-	### Returns:
-	 - `ParamsDict` 
-	   modified copy of `params_base`
-	"""
-
-	# copy the input dict
-	output : dict = copy.deepcopy(params_base)
-
-	# REVIEW: why did i even refactor this when im making everything editable through params json anyway?
-	for tup_key,val in params_mod.items():
-		# merge in the standard params
-		if tup_key.mod_type == ModTypes.params.value:
-			
-			nested_keys : str = tup_key.path
-
-			fin_dic,fin_key = keylist_access_nested_dict(
-				d = output, 
-				keys = nested_keys.split('.'),
-			)
-			fin_dic[fin_key] = val
-
-		elif tup_key.mod_type == ModTypes.conn.value:
-			# merge in the connection modifiers
-			conn_key_str : str = tup_key.path
-			conn_key = strList_to_dict(
-				in_data = conn_key_str,
-				keys_list = ['NS', 'from', 'to', 'type'],
-				delim = ',',
-			)
-
-			# get the indecies of the connections whose weights need to be changed
-			conn_idxs : List[Optional[int]] = find_conn_idx_regex(
-				params_data = output, 
-				conn_key = conn_key,
-			)
-
-			# set weights
-			for cidx in conn_idxs:
-				output[conn_key['NS']]['connections'][cidx]['weight'] = val
-		else:
-			raise NotImplementedError(f'given key type {tup_key.mod_type} unknown')
-
-	return output
-
-
-def extract_mods_from_params(
-		params : ParamsDict, 
-		modkeys : List[ModParam],
-		default_val : Any = None,
-	) -> Tuple[ParamsDict,ModParamsDict]:
-	"""extracts modparams from a params dict (inverts `merge_params_with_mods`)
-
-	returns `Tuple[ParamsDict,ModParamsDict]` containing:
-	- copy of the params dict, with extracted values overwritten with `default_val`
-	- the mod params dict
-
-	combining these two using `merge_params_with_mods` should just give back `params`
-	"""
-
-	params_stripped : ParamsDict = deepcopy(params)
-	modparams : ModParamsDict = dict()
-
-	for key in modkeys:
-		# merge in the standard params
-		if key.mod_type == ModTypes.params.value:
-			# access the element
-			fin_dic,fin_key = keylist_access_nested_dict(
-				d = params_stripped, 
-				keys = key.path.split('.'),
-			)
-			# store copy in modparams
-			modparams[key] = deepcopy(fin_dic[fin_key])
-			# overwrite in params_stripped
-			fin_dic[fin_key] = default_val
-
-		elif key.mod_type == ModTypes.conn.value:
-			# merge in the connection modifiers
-
-			# get the key
-			conn_key = strList_to_dict(
-				in_data = key.path,
-				keys_list = ['NS', 'from', 'to', 'type'],
-				delim = ',',
-			)
-
-			# get the indecies of the connections whose weights need to be changed
-			conn_idxs : List[Optional[int]] = find_conn_idx_regex(
-				params_data = params_stripped,
-				conn_key = conn_key,
-			)
-
-			for cidx in conn_idxs:
-				# store weights in mod params
-				modparams[key] = (
-					params_stripped
-					[conn_key['NS']]
-					['connections']
-					[cidx]
-					['weight']
-				)
-
-				params_stripped[conn_key['NS']]['connections'][cidx]['weight'] = default_val
-
-		else:
-			raise NotImplementedError(f'given key type {key.mod_type} unknown')
-
-	return params_stripped,modparams
-
-
-
-def jointo_nan_eval_runs(eval_runs : List[ModParamsDict]) -> ModParamsDict:
-	"""
-	create a `ModParamsDict` that maps every key from `eval_runs` to float('nan')
-	
-	useful for then merging into params
-	"""
-	joined_nans : ModParamsDict = dict()
-
-	for x in eval_runs:
-		for k in x.keys():
-			joined_nans[k] = float('nan')
-
-	return joined_nans
-
-	
-
-
-
-"""
-
- ###### #    # ##### #####    ##    ####  #####
- #       #  #    #   #    #  #  #  #    #   #
- #####    ##     #   #    # #    # #        #
- #        ##     #   #####  ###### #        #
- #       #  #    #   #   #  #    # #    #   #
- ###### #    #   #   #    # #    #  ####    #
-
-"""
-
-ExtractorReturnType = Any
-
-ExtractorFunc = Callable[
-	[
-		Arg(Path, 'datadir'),
-		Arg(ParamsDict, 'params'),
-		Arg(bool, 'ret_nan'),
-	], 
-	ExtractorReturnType, # return type
-]
-
-MultiExtractorFunc = Callable[
-	[
-		Arg(Path, 'datadir'),
-		Arg(ParamsDict, 'params'),
-		Arg(bool, 'ret_nan'),
-	], 
-	float, # return type
-]
-
-
-
-def _wrapper_extract(
-		proc, 
-		func_extract : ExtractorFunc, 
-		outpath : Path, 
-		params_joined : ParamsDict,
-	):	
-	# wait for command to finish
-	proc.wait()
-		
-	if proc.returncode:
-		print(f'  >>  ERROR: process terminated with exit code 1, check log.txt for:\n        {str(proc.args)}')
-
-	result : ExtractorReturnType = func_extract(
-		datadir = outpath,
-		params = params_joined,
-		ret_nan = bool(proc.returncode),
-	)
-
-	with open(joinPath(outpath, 'extracted.txt'), 'a') as fout_ext:
-		print(f'# extracted using {func_extract.__name__}:', file = fout_ext)
-		print(repr(result), file = fout_ext)
-	
-	return result
-
-
-def _wrap_multi_extract(
-		func_extract : ExtractorFunc,
-		calc_mean : Callable[[List[float]], float] = lambda x : min(x),
-	) -> MultiExtractorFunc:
-
-	def _func_extract_MULTI(
-			datadir : Path,
-			params : ParamsDict,
-			ret_nan : bool = False,
-		) -> float:
-
-		# TODO: ret_nan is not actually doing the correct thing here, although its probably unimportant. current implementation does not allow for just one of several processes failing
-
-		lst_extracted : List[float] = list()
-
-		for p in os.listdir(datadir):
-			p_joined : Path = joinPath(datadir,p)
-			if os.path.isdir(p_joined):
-				lst_extracted.append(func_extract(
-					datadir = p_joined,
-					params = params,
-					ret_nan = ret_nan,
-				))
-
-		return calc_mean(lst_extracted)
-
-	# add metadata
-	_func_extract_MULTI.__name__ = func_extract.__name__
-	_func_extract_MULTI.__doc__ = f"""
-		{_wrap_multi_extract.__doc__} 
-		#### docstring of wrapped function:
-		```markdown
-		{func_extract.__doc__}
-		```
-	"""
-	return _func_extract_MULTI
-
-
-def _extract_TEMPLATE(
-		datadir : Path,
-		params : ParamsDict,
-		ret_nan : bool = False,
-	) -> ExtractorReturnType:
-	"""template function for extraction functions
-	
-	dont actually call this function. it contains documentation for the format of functions 
-	`func_extract` taken by `evaluate_params()`
-	
-	### Parameters:
-	 - `datadir : Path`   
-	   output directory of data
-	 - `params : ParamsDict`   
-	   nested dictionary of params
-	 - `ret_nan : bool`   
-	   whether to return nan value (when process terminates in error)
-	   (defaults to `False`)
-	
-	### Returns:
-	 - `ExtractorReturnType` 
-	   can return any data about the run
-	
-	### Raises:
-	 - `NotImplementedError` : dont run this!
-	"""
-	
-	raise NotImplementedError('this is a template function only!')
-
-
-def extract_finalpos(
-		datadir : Path,
-		params : ParamsDict,
-		ret_nan : bool = False,
-	) -> Tuple[float,float]:
-	"""extract just the final head position
-	
-	### Returns:
-	 - `Tuple[float,float]` 
-	   head position
-	"""
-	if ret_nan:
-		return ( float('nan'), float('nan') )	
-	else: 
-		bodydata : CoordsRotArr = read_body_data(joinPath(datadir,'body.dat'))[-1,0]
-		return ( bodydata['x'], bodydata['y'] )
-
-def extract_food_dist(
-		datadir : Path,
-		params : ParamsDict,
-		ret_nan : bool = False,
-	) -> float:
-	"""extract euclidean distance from head to food
-	
-	### Returns:
-	 - `float` 
-	   dist from final head position to food
-	"""
-	if ret_nan:
-		return float('nan')
-	else:
-		# get head pos
-		bodydata : CoordsRotArr = read_body_data(joinPath(datadir,'body.dat'))[-1,0]
-		pos_head : VecXY = VecXY( bodydata['x'], bodydata['y'] )
-
-		# get food pos
-		pos_food : VecXY = VecXY(
-			params['ChemoReceptors']['foodPos']['x'],
-			params['ChemoReceptors']['foodPos']['y'],
-		)
-
-		# return distance
-		return dist(pos_head, pos_food)
-
-def extract_food_dist_inv(
-		datadir : Path,
-		params : ParamsDict,
-		ret_nan : bool = False,
-	) -> float:
-	"""extract inverse of euclidean distance from head to food
-
-	this means that higher value ==> higher fitness
-	
-	### Returns:
-	 - `float` 
-	   1/(dist from final head position to food)
-	"""
-	return 1 / extract_food_dist(datadir, params, ret_nan)
-
-
-def extract_df_row(
-		datadir : Path,
-		params : ParamsDict,
-		ret_nan : bool = False,
-	) -> dict:
-	# TODO: implement extracting more data, for parameter sweeps
-	raise NotImplementedError('please implement me :(')
-
 
 """
