+++ pyutil/util.py
@@ -249,215 +249,6 @@ def dict_hash(data : dict, hash_len_mod : int = int(10**8)) -> int:
 	return hash(tuple(data.items())) % hash_len_mod
 
 
-"""
-
- #    #  ####  #####     #####    ##   #####    ##   #    #  ####
- ##  ## #    # #    #    #    #  #  #  #    #  #  #  ##  ## #
- # ## # #    # #    #    #    # #    # #    # #    # # ## #  ####
- #    # #    # #    #    #####  ###### #####  ###### #    #      #
- #    # #    # #    #    #      #    # #   #  #    # #    # #    #
- #    #  ####  #####     #      #    # #    # #    # #    #  ####
-
-"""
-
-Valid_NS = Literal['Head', 'VentralCord']
-Valid_Neurons = str
-# Valid_Neurons = Literal['Head', 'VentralCord']
-
-class ModTypes(Enum):
-	params : str = 'params'
-	conn : str = 'conn'
-	# cli : str = 'cli'
-
-# T_ModTypes = Literal[tuple(e.value for e in ModTypes)]
-T_ModTypes = Literal['params', 'conn']
-
-# ModTypes = Literal[
-# 	'params',
-# 	'conn',
-# 	'cli',
-# ]
-
-ModParam = NamedTuple(
-	'ModParam', 
-	[
-		('mod_type', T_ModTypes), 
-		('path', str),
-	],
-)
-
-RangeTuple = NamedTuple(
-	'RangeTuple', 
-	[
-		('min', float), 
-		('max', float),
-	],
-)
-
-NormalDistTuple = NamedTuple(
-	'NormalDistTuple', 
-	[
-		('mu', float), 
-		('sigma', float),
-	],
-)
-
-DistTuple = Union[RangeTuple,NormalDistTuple]
-
-
-
-ParamsDict = Dict[str, Any]
-ModParamsDict = Dict[ModParam, float]
-ModParamsRanges = Dict[ModParam, RangeTuple]
-ModParamsDists = Dict[ModParam, DistTuple]
-
-def distributions_to_ranges(in_data : ModParamsDists, n_sigma : float = 1.5) -> ModParamsRanges:
-	output : ModParamsRanges = dict()
-	for k,v in in_data.items():
-		if isinstance(v, RangeTuple):
-			output[k] = v
-		elif isinstance(v, NormalDistTuple):
-			output[k] = RangeTuple(
-				min = v.mu - v.sigma * n_sigma,
-				max = v.mu + v.sigma * n_sigma,
-			)
-		else:
-			raise NotImplementedError(f"unknown distribution type:\t{k}\t{v}\t{type(v)}")
-	
-	return output
-
-
-
-def load_params(path : Path) -> ParamsDict:
-	with open(path, 'r') as fin:
-		data : ParamsDict = json.load(fin)
-	return data
-
-
-# TODO: delete this function
-def modprmdict_to_filename(
-		data : ModParamsDict,
-		key_order : Optional[List[ModParam]] = None,
-		short_keys : Optional[int] = None,
-	):
-	
-	if key_order is None:
-		key_order = list(data.keys())
-
-	output : List[str] = []
-
-	for k in key_order:
-		k_write : str = str(k)
-		
-		# UGLY: this whole bit
-		if k.mod_type == ModTypes.conn.value:
-			_,str_from,str_to,_ = k.path.split(',')
-			k_write = f"{str_from}-{str_to}".replace('*','x')
-		elif k.mod_type == ModTypes.params.value:
-			k_write = k.path.split('.')[-1][:short_keys]
-		
-		output.append(f'{k_write}={data[k]:.3}')
-	
-	return '_'.join(output)
-
-# ConnKey = NamedTuple(
-# 	'ConnKey',
-# 	[
-# 		('NS', Valid_NS),
-# 		('from', Valid_Neurons),
-# 		('to', Valid_Neurons),
-# 		('weight', float),
-# 	],
-# )
-
-
-"""
- ######   #######  ##    ## ##    ##
-##    ## ##     ## ###   ## ###   ##
-##       ##     ## ####  ## ####  ##
-##       ##     ## ## ## ## ## ## ##
-##       ##     ## ##  #### ##  ####
-##    ## ##     ## ##   ### ##   ###
- ######   #######  ##    ## ##    ##
-"""
-
-def find_conn_idx(params_data : Dict[str,Any], conn_key : dict) -> Optional[int]:
-	"""finds the index of the entry matching conn_key"""
-
-	for i,item in enumerate(params_data):
-		if all([
-				conn_key[k] == item[k]
-				for k in conn_key 
-				if k != 'NS'
-			]):
-			return i
-
-	return None
-
-
-def find_conn_idx_regex(
-		params_data : Dict[str,Any], 
-		conn_key : dict,
-		# special_scaling_map : Optional[Dict[str,float]] = None,
-	) -> List[Optional[int]]:
-
-	conn_idxs : List[Optional[int]] = [None]
-
-	# UGLY: clean this bit up
-
-	if conn_key['to'].endswith('*'):
-		# if wildcard given, find every connection that matches
-		conn_idxs = list()
-		
-		conn_key_temp : dict = deepcopy(conn_key)
-		
-		for nrn in params_data[conn_key['NS']]['neurons']:
-			# loop over neuron names, check if they match
-			# REVIEW: this isnt full regex, but whatever
-			if nrn.startswith(conn_key['to'].split('*')[0]):
-				conn_key_temp['to'] = nrn
-				cidx_temp : Optional[int] = find_conn_idx(
-					params_data[conn_key_temp['NS']]['connections'],
-					conn_key_temp,
-				)
-				# append to list, but only if an existing connection is found
-				# note that this behavior differs from when no wildcard is given,
-				# in that new connections will not be created
-				if cidx_temp is not None:
-					conn_idxs.append(cidx_temp)
-	elif conn_key['from'].endswith('*'):
-		# if wildcard given, find every connection that matches
-		conn_idxs = list()
-		
-		conn_key_temp : dict = deepcopy(conn_key)
-		
-		for nrn in params_data[conn_key['NS']]['neurons']:
-			# loop over neuron names, check if they match
-			# REVIEW: this isnt full regex, but whatever
-			if nrn.startswith(conn_key['from'].split('*')[0]):
-				conn_key_temp['from'] = nrn
-				cidx_temp : Optional[int] = find_conn_idx(
-					params_data[conn_key_temp['NS']]['connections'],
-					conn_key_temp,
-				)
-				# append to list, but only if an existing connection is found
-				# note that this behavior differs from when no wildcard is given,
-				# in that new connections will not be created
-				if cidx_temp is not None:
-					conn_idxs.append(cidx_temp)
-	else:
-		# if special_scaling_map is not None:
-		# 	raise ValueError(f"`special_scaling_map` specified, but no wildcard given in neuron name:   {special_scaling_map}    {conn_key['to']}")
-
-		# if no wildcard specified, just get the one connection
-		conn_idxs = [ find_conn_idx(
-			params_data[conn_key['NS']]['connections'],
-			conn_key,
-		) ]
-	
-	return conn_idxs
-
-
 
 """
  ######  ##     ## ########
